Update Redis and Redlock

+++ src/MassTransit.RedisSagas.RedLock/RedisSagaConsumeContext.cs
@@ -4,24 +4,19 @@ using MassTransit.Context;
 using MassTransit.Logging;
 using MassTransit.RedisSagas.RedLock;
 using MassTransit.Util;
-using RedLock;
+using RedLockNet;
 using StackExchange.Redis;
 
 namespace MassTransit.RedisSagas
 {
-    public class RedLockSagaConsumeContext<TSaga, TMessage> :
-        ConsumeContextProxyScope<TMessage>,
-        SagaConsumeContext<TSaga, TMessage>
-        where TMessage : class
-        where TSaga : class, IVersionedSaga
+    public class RedLockSagaConsumeContext<TSaga, TMessage> : ConsumeContextProxyScope<TMessage>, SagaConsumeContext<TSaga, TMessage> where TMessage : class where TSaga : class, IVersionedSaga
     {
-        static readonly ILog Log = Logger.Get<RedLockSagaRepository<TSaga>>();
-        readonly IDatabase _redisDb;
-        private readonly IRedisLockFactory _lockFactory;
+        private static readonly ILog Log = Logger.Get<RedLockSagaRepository<TSaga>>();
+        private readonly IDistributedLockFactory _lockFactory;
+        private readonly IDatabase _redisDb;
         private readonly string _redisPrefix;
 
-        public RedLockSagaConsumeContext(IDatabase redisDb, IRedisLockFactory lockFactory, ConsumeContext<TMessage> context, TSaga instance, string redisPrefix = "")
-            : base(context)
+        public RedLockSagaConsumeContext(IDatabase redisDb, IDistributedLockFactory lockFactory, ConsumeContext<TMessage> context, TSaga instance, string redisPrefix = "") : base(context)
         {
             Saga = instance;
             _redisDb = redisDb;
@@ -41,17 +36,14 @@ namespace MassTransit.RedisSagas
 
         async Task SagaConsumeContext<TSaga>.SetCompleted()
         {
-            ITypedDatabase<TSaga> db = _redisDb.As<TSaga>();
+            var db = _redisDb.As<TSaga>();
 
-            using(var distLock  = _lockFactory.Create($"redislock:{Saga.CorrelationId}", TimeSpan.FromSeconds(5), TimeSpan.FromSeconds(5), TimeSpan.FromSeconds(0.5)))
+            using (var distLock = await _lockFactory.CreateLockAsync($"redislock:{Saga.CorrelationId}", TimeSpan.FromSeconds(5), TimeSpan.FromSeconds(5), TimeSpan.FromSeconds(0.5)))
             {
                 if (Log.IsDebugEnabled)
                     Log.Debug($"SAGA:{TypeMetadataCache<TSaga>.ShortName}:{TypeMetadataCache<TMessage>.ShortName} Entering Lock {Saga.CorrelationId}");
 
-                if (distLock.IsAcquired)
-                {
-                    await db.Delete(Saga.CorrelationId, _redisPrefix).ConfigureAwait(false);
-                }
+                if (distLock.IsAcquired) await db.Delete(Saga.CorrelationId, _redisPrefix).ConfigureAwait(false);
 
                 if (Log.IsDebugEnabled)
                     Log.Debug($"SAGA:{TypeMetadataCache<TSaga>.ShortName}:{TypeMetadataCache<TMessage>.ShortName} Leaving Lock {Saga.CorrelationId}");
